refactor: minor names changes, move core urls, views and forms to app administrativo
