+++ apps/clientes/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ClientesConfig(AppConfig):
+class ClientsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'apps.clientes'
