+++ apps/clientes/forms.py
@@ -1,57 +1,8 @@
 from django import forms
-from apps.administrativo.models import Enterprise, Card, Locker
 from apps.clientes.models import Client
 from apps.core.models import Address
 
-class LockerForm(forms.Form):
-    available = forms.BooleanField(
-        required=False,
-        initial=True,
-        label="Disponibilidade pra uso",
-        widget=forms.CheckboxInput(attrs={'class': 'form-check-input'})
-    )  
-    number = forms.IntegerField(
-        label="Número do Armário",
-        min_value=0, 
-        widget=forms.NumberInput(attrs={'class': 'form-control', 'placeholder': 'Digite o nº do armário'})
-    )
-    card = forms.ModelChoiceField(
-        required=False,
-        queryset=Card.objects.all(),
-        label="Atribuir Cartão Chave",
-        widget=forms.Select(attrs={'class': 'form-select'})
-    )
-    enterprise = forms.ModelChoiceField(
-        queryset=Enterprise.objects.all(),
-        label="Empresa:",
-        widget=forms.Select(attrs={'class': 'form-select'})
-    )
-class LockerAssignmentForm(forms.Form):
-    available = forms.BooleanField(
-        required=False, 
-        initial=True, 
-        label="Disponibilidade pra uso",
-        widget=forms.CheckboxInput(attrs={'class': 'form-check-input'})
-    )  
-    number = forms.IntegerField(
-        required=False,
-        label="Número do Armário",
-        min_value=1, 
-        widget=forms.NumberInput(attrs={'class': 'form-control', 'disabled': 'disabled', 'readonly': 'readonly'})
-    )
-    card = forms.ModelChoiceField(
-        queryset=Card.objects.all(),
-        required=False,
-        label="Atribuir Cartão Chave",
-        widget=forms.Select(attrs={'class': 'form-select', 'disabled': 'disabled', 'readonly': 'readonly'})
-    )
-    client = forms.ModelChoiceField(
-        queryset=Client.objects.all(),
-        label="Atribuir Usuário:",
-        widget=forms.Select(attrs={'class': 'form-select'})
-    )
-
-class UserForm(forms.Form):
+class ClientForm(forms.Form):
     name = forms.CharField(
         max_length=255,
         label="Nome Completo:",
@@ -114,4 +65,3 @@ class UserForm(forms.Form):
         label="Ponto de referência",        
         widget=forms.TextInput(attrs={'class': 'form-control'})
     )
-
