+++ apps/clientes/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [  
+    path('clients/', views.clients, name='clients'),     
+    path('client/add/', views.add_client, name='add_client'), 
+]
