+++ lockin/urls.py
@@ -20,5 +20,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('apps.core.urls')),
+    path('', include('apps.administrativo.urls')),
+    path('', include('apps.clientes.urls')),
 ]
