+++ static/css/base/dataTables.custom.css
@@ -17,7 +17,8 @@ table.dataTable thead > tr > td.dt-orderable-asc:hover,
 table.dataTable thead > tr > td.dt-orderable-desc:hover {
     outline: 0;
     outline-offset: 0;
-    background-color: #dcf9df;
+    color: #1B84FF;
+    background-color: #F1FAFF;
 }
 
 table.dataTable th.dt-type-numeric,
