+++ templates/client/add_client.html
@@ -3,7 +3,7 @@
 {% block title %}LockIn - Cadastrar Usuário{% endblock %}
 
 {% block content %}
-<!--begin::Add User-->
+<!--begin::Add Client-->
 <main class="col-md-9 ms-sm-auto col-lg-10 px-md-4 mt-3 mb-3 mt-lg-5 mb-lg-5">
     <div class="card p-3 p-lg-5">
         <div class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-4 mb-lg-5 border-bottom">
@@ -11,7 +11,7 @@
         </div>
         <form method="POST">
             {% csrf_token %}               
-            {% include "form/alert_errors.html" %}
+            {% include "partials/forms/alert_errors.html" %}
 
             <div class="border rounded p-3 p-lg-5">             
                 <div class="row mb-3">
@@ -78,5 +78,5 @@
           </form>
     </div>
 </main>
-<!--end::Add User-->
+<!--end::Add Client-->
 {% endblock %}
