+++ templates/locker/add_locker.html
@@ -12,7 +12,7 @@
     </div>
     <form method="POST">
       {% csrf_token %}
-      {% include "form/alert_errors.html" %}
+      {% include "partials/forms/alert_errors.html" %}
 
       <div class="border rounded p-3 p-lg-5">
         <div class="row mb-3">
