+++ templates/partials/forms/messages.html
@@ -1,0 +1,7 @@
+{% if messages %}
+{% for message in messages %}
+<div class="alert alert-{{ message.tags }}">
+  {{ message }}
+</div>
+{% endfor %}
+{% endif %}
