--- templates/user/users.html
@@ -1,48 +1,0 @@
-{% extends 'base.html' %}
-
-{% block title %}LockIn - Usuários{% endblock %}
-
-{% block content %}
-<!--begin::User List-->
-<main class="col-md-9 ms-sm-auto col-lg-10 px-md-4 mt-3 mb-3 mt-lg-5 mb-lg-5">
-  <div class="card p-3 p-lg-5">
-    <div
-      class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-4 mb-lg-5 border-bottom">
-      <h1 class="h2">Usuários</h1>
-    </div>
-
-    <!-- Campo de busca -->
-    <div class="mb-3">
-      <input type="text" id="busca-armario" class="form-control"
-        placeholder="Buscar armário por nome ou localização...">
-    </div>
-
-    <!-- Tabela -->
-    <table class="table table-striped table-hover" id="tabela-armarios">
-      <thead class="table-dark">
-        <tr>
-          <th>#</th>
-          <th>Nome</th>
-          <th>Telefone</th>
-          <th>E-mail</th>
-        </tr>
-      </thead>
-      <tbody>
-        {% for user in users %}
-        <tr>
-          <td>{{ forloop.counter }}</td>
-          <td>{{ user.name }}</td>
-          <td>{{ user.phone }}</td>      
-          <td>{{ user.email }}</td>        
-        </tr>
-        {% empty %}
-        <tr>
-          <td colspan="5" class="text-center text-muted">Nenhum usuário cadastrado.</td>
-        </tr>
-        {% endfor %}
-      </tbody>
-    </table>
-  </div>
-</main>
-<!--end::User List-->
-{% endblock %}
